polyglot-backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polyglot-backend"
version = "0.1.0"
description = "Backend for Polyglot Meeting Whisperer"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Polyglot Team" },
]

dependencies = [
    "gradio>=4.19.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "pydub>=0.25.1",
    "aiohttp>=3.8.5",
    "openai>=1.13.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
polyglot-backend = "main:app"

[tool.hatch.build.targets.wheel]
packages = ["main.py"]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=7.0.0",
]

[tool.hatch.envs.dev]
dependencies = [
    "black>=23.1.0",
    "isort>=5.12.0",
    "mypy>=1.0.1",
] 